src/gateway/requirements.txt
fastapi
google-auth
Jinja2
uvicorn  # starlette inside
